Update and rename UnsafecsrfComparison.java to csrfComparison.java
